Bugs resolvidos
