+++ .env.example
@@ -1,0 +1,13 @@
+MYSQL_HOST=
+MYSQL_DATABASE=
+MYSQL_PASSWORD=
+MYSQL_ROOT_PASSWORD=
+MYSQL_ROOT_USER=
+MYSQL_PORT=
+
+NODEJS_PORT=
+
+JWT_SECRET=
+
+MAIL_USER=
+MAIL_PASSWORD=
