+++ controllers/connectController.js
@@ -1,0 +1,3 @@
+exports.connection = (req, res) => {
+  return res.status(200).json({ status: true });
+};
