--- marvie_backend/marvie_backend/controllers/connectController.js
@@ -1,3 +1,0 @@
-exports.connection = (req, res) => {
-  return res.status(200).json({ status: true });
-};
