--- marvie_backend/marvie_backend/routes/connectRoutes.js
@@ -1,8 +1,0 @@
-const express = require("express");
-const router = express.Router();
-const connectController = require("../controllers/connectController");
-const verifyJWT = require("../config/auth");
-
-router.get("/", connectController.connection);
-
-module.exports = router;
