--- marvie_backend/marvie_backend/routes/ordersRoutes.js
@@ -1,9 +1,0 @@
-const express = require("express");
-const router = express.Router();
-const ordersController = require("../controllers/ordersController");
-const verifyJWT = require("../config/auth");
-
-router.get("/", verifyJWT, ordersController.getOrders);
-router.get("/:id", verifyJWT, ordersController.getOrdersById);
-
-module.exports = router;
