--- marvie_backend/marvie_backend/routes/tokenRoutes.js
@@ -1,8 +1,0 @@
-const express = require("express");
-const router = express.Router();
-const tokenController = require("../controllers/tokenController");
-const verifyJWT = require("../config/auth");
-
-router.get("/", verifyJWT, tokenController.verifyToken);
-
-module.exports = router;
