+++ routes/passwordResetRoutes.js
@@ -1,0 +1,9 @@
+const express = require("express");
+const router = express.Router();
+const passwordResetController = require("../controllers/passwordResetController");
+
+router.post("/request", passwordResetController.requestPasswordReset);
+
+router.post("/reset", passwordResetController.resetPassword);
+
+module.exports = router;
